babu v.2.0/requirements.txt
Flask
requests
SpeechRecognition
pydub
google-generativeai
Pillow
edge-tts
python-dotenv
sounddevice
soundfile
numpy
schedule
opencv-python
